Add type hints and lint
