+++ .isort.cfg
@@ -1,7 +1,6 @@
 [settings]
 default_section=THIRDPARTY
 indent='    '
-known_first_party=request_profiler
 sections=FUTURE,STDLIB,THIRDPARTY,FIRSTPARTY,LOCALFOLDER
 multi_line_output=3
 line_length=88
