+++ .pre-commit-config.yaml
@@ -1,14 +1,50 @@
 repos:
-    - repo: https://github.com/pre-commit/mirrors-pylint
-      rev: v2.3.1
-      hooks:
-          - id: pylint
+    # python import sorting - will amend files
     - repo: https://github.com/pre-commit/mirrors-isort
       rev: v4.3.21
       hooks:
           - id: isort
+
+    # python code formatting - will amend files
     - repo: https://github.com/ambv/black
-      rev: 19.3b0
+      rev: 19.10b0
       hooks:
           - id: black
-            language_version: python3.7
+
+    # PEP8 linting, with added Django goodness, and custom YJ plugin
+    - repo: https://github.com/pre-commit/mirrors-pylint
+      rev: v2.4.4
+      hooks:
+          - id: pylint
+            args:
+                - --rcfile=.pylintrc
+
+    # Flake8 includes pyflakes, pycodestyle, mccabe, pydocstyle, bandit
+    - repo: https://github.com/pre-commit/pre-commit-hooks
+      rev: v2.4.0
+      hooks:
+          - id: flake8
+            additional_dependencies: ["flake8-docstrings", "flake8-bandit"]
+            exclude: tests
+
+    # python static type checking
+    - repo: https://github.com/pre-commit/mirrors-mypy
+      rev: v0.761
+      hooks:
+          - id: mypy
+            args:
+                - --disallow-untyped-defs
+                - --disallow-incomplete-defs
+                - --check-untyped-defs
+                - --no-implicit-optional
+                - --ignore-missing-imports
+                - --follow-imports=silent
+            exclude: ^tests
+
+    # JS, JSON and SCSS formatting
+    - repo: https://github.com/prettier/prettier
+      rev: 1.19.1
+      hooks:
+          - id: prettier
+            args:
+                - --config=.prettierrc
