+++ .travis.yml
@@ -1,16 +1,6 @@
 language: python
 python:
     - "3.7"
-
-before_install:
-    - curl -sSL https://raw.githubusercontent.com/sdispater/poetry/master/get-poetry.py | python
-    - source $HOME/.poetry/env
-
-install:
-    - poetry install -v
-    - pip install tox-travis
-
-script:
-    - poetry run isort --check-only --recursive command_log
-    - poetry run black --check command_log
-    - poetry run tox
+    - "3.8"
+install: pip install tox-travis
+script: tox
