+++ CONTRIBUTING.md
@@ -1,26 +1,24 @@
 # Contributing
 
-We, as YunoJuno, will be building this app out to support our specific
-requirements. New features will be added at our convenience.
+We, as YunoJuno, will be building this app out to support our specific requirements. New features
+will be added at our convenience.
 
-All of which means, if the app isn't currently supporting _your_ use case,
-**get involved**!
+All of which means, if the app isn't currently supporting _your_ use case, **get involved**!
 
 The usual rules apply:
 
-1. If you see something that's wrong, or something that's missing, open an
-issue. NB please take a minute to verify that your issue is not already
-covered by an existing issue.
+1. If you see something that's wrong, or something that's missing, open an issue. NB please take a
+   minute to verify that your issue is not already covered by an existing issue.
 
 2. If you want to fix something, or add a new feature / backend etc. then:
 
-* Fork the repo
-* Run the tests locally: `python manage.py test test_app request_profiler`
-* Create a local branch (ideally named after the related issue,
-  beginning with the issue number, e.g. `1-add-template-timings`)
-* Write some code
-* Write some tests to prove your code works
-* Commit it
-* Send a pull request
+-   Fork the repo
+-   Run the tests locally: `python manage.py test test_app request_profiler`
+-   Create a local branch (ideally named after the related issue, beginning with the issue number,
+    e.g. `1-add-template-timings`)
+-   Write some code
+-   Write some tests to prove your code works
+-   Commit it
+-   Send a pull request
 
 Other than that we'll work it out as we go along.
