+++ LICENSE
@@ -1,6 +1,6 @@
 The MIT Licence (MIT)
 
-Copyright (c) 2014 YunoJuno Ltd
+Copyright (c) 2020 YunoJuno Ltd
 
 Permission is hereby granted, free of charge, to any person obtaining a copy
 of this software and associated documentation files (the "Software"), to deal
