+++ mypy.ini
@@ -1,0 +1,21 @@
+[mypy]
+strict_optional=True
+ignore_missing_imports=True
+follow_imports=silent
+warn_redundant_casts=True
+warn_unused_ignores = true
+warn_unreachable = true
+disallow_untyped_defs = true
+disallow_incomplete_defs = true
+
+# Disable mypy for migrations
+[mypy-*.migrations.*]
+ignore_errors=True
+
+# Disable mypy for settings
+[mypy-*.settings.*]
+ignore_errors=True
+
+# Disable mypy for tests
+[mypy-*.tests.*]
+ignore_errors=True
