+++ pytest.ini
@@ -1,3 +1,3 @@
 [pytest]
-DJANGO_SETTINGS_MODULE = test_app.settings
-python_files = tests.py
+DJANGO_SETTINGS_MODULE = tests.settings
+python_files = test_*.py
