+++ request_profiler/apps.py
@@ -5,4 +5,3 @@ class RequestProfilerAppConfig(AppConfig):
 
     name = "request_profiler"
     verbose_name = "Request Profiler"
-    configs = []
