--- test_app/templates/test.html
@@ -1,8 +1,0 @@
-<html>
-<head>
-<title>Test</title>
-</head>
-<body>
-This is a test.
-</body>
-</html>
