+++ tests/templates/test.html
@@ -1,0 +1,8 @@
+<html>
+    <head>
+        <title>Test</title>
+    </head>
+    <body>
+        This is a test.
+    </body>
+</html>
